requirements.txt
requests==2.31.0
Flask==3.0.0
PyYAML==6.0.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0
coverage==7.3.2